Move NotificationConfig object to the config module.
